Update frequencies to 868.x only, default external antenna
The old version used the 868.x and 864.x frequencies as default when
joining the network. The new version of the spec uses 868.x
exclusively. Most gateways seems to be using these frequencies.

The default for the EE-02 antenna was to use the internal antenna.
In the new version the chip antenna is removed and replaced by a
soldering connection. The uFL connector should be the default.